requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
urllib3==2.0.7